fix: 将滑块组件的浮点参数改为整型参数
避免浮点数精度问题，确保滑块组件参数处理的一致性

+++ src/main/java/cn/ningmo/geysermenu/MenuManager.java
@@ -316,10 +316,10 @@ public class MenuManager {
                     }
                     
                     case "slider" -> {
-                        float min = getFloat(component, "min", 0);
-                        float max = getFloat(component, "max", 100);
-                        float step = getFloat(component, "step", 1);
-                        float defaultVal = getFloat(component, "default", min);
+                        int min = getInt(component, "min", 0);
+                        int max = getInt(component, "max", 100);
+                        int step = getInt(component, "step", 1);
+                        int defaultVal = getInt(component, "default", min);
                         form.slider(text, min, max, step, defaultVal);
                     }
